backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "asgiref (>=3.9.1,<4.0.0)",
    "django (>=5.2.5,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-stubs-ext (>=5.2.2,<6.0.0)",
    "django-rest-passwordreset (>=1.5.0,<2.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-factoryboy (>=2.8.1,<3.0.0)",
    "psutil (>=7.2.1,<8.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
